Upload 365, find, doramatv. Add w8 selector in ADD image VK video. Update updater. 0.1.2.2

+++ autorization/checker_autorization.py
@@ -3,6 +3,7 @@ from autorization.autorization_animaunt.autorization_web_animaunt import Animaun
 from autorization.autorization_malfurik.autorization_web_malfurik import Malfurik_web
 from PyQt5.QtCore import QThread, pyqtSignal
 import os
+from playwright.sync_api import sync_playwright
 
 
 class CheckerThread(QThread):
@@ -10,26 +11,51 @@ class CheckerThread(QThread):
     finished = pyqtSignal(bool, bool, bool)
 
     def run(self):
-        try:
+        vk_check = False
+        ani_check = False
+        malf_check = False
+
+        with sync_playwright() as pl:
+            browser = pl.chromium.launch(headless=True)
+
             if os.path.exists("./assets/vk_storage.json"):
-                vk_check = AutorizationWebVK(checker=True).check_autorezade(link="https://vk.com/", checker=True)
-            else:
-                vk_check = False
-            
+                try:
+                    context = browser.new_context(storage_state="assets/vk_storage.json")
+                    page = context.new_page()
+                    page.goto("https://vk.com/")
+                    if page.url.format() != "https://vk.com/":
+                        vk_check = True
+                    else:
+                        os.remove("./assets/vk_storage.json")
+                    context.close()
+                except:
+                    pass
             if os.path.exists("./assets/animaunt_storage.json"):
-                ani_check = Animaunt_web(True).checker()
-            else:
-                ani_check = False
-            
+                try:
+                    context = browser.new_context(storage_state="assets/animaunt_storage.json")
+                    page = context.new_page()
+                    page.goto(os.getenv('ANIMAUNT_LINK'))
+                    try:
+                        page.wait_for_selector(".sidebar-user-material", timeout=20000)
+                        ani_check = True
+                    except:
+                        os.remove("./assets/animaunt_storage.json")
+                    context.close()
+                except:
+                    pass
             if os.path.exists("./assets/malfurik_storage.json"):
-                malf_check = Malfurik_web(True).checker()
-            else:
-                malf_check = False
-            # vk_check, malf_check, ani_check = True, True, True
-            # Отправка результата в основной поток через сигнал
-            self.finished.emit(malf_check, ani_check, vk_check)
-        except Exception as e:
-            os.remove("./assets/vk_storage.json")
-            os.remove("./assets/animaunt_storage.json")
-            os.remove("./assets/malfurik_storage.json")
-            self.finished.emit(False, False, False)
+                try:
+                    context = browser.new_context(storage_state="assets/malfurik_storage.json")
+                    page = context.new_page()
+                    page.goto("https://anime.malfurik.online/")
+                    wpadminbar_element = page.query_selector('#wpadminbar')
+                    if wpadminbar_element:
+                        malf_check = True
+                    else:
+                        os.remove("./assets/malfurik_storage.json")
+                    context.close()
+                except:
+                    pass
+            browser.close()
+
+        self.finished.emit(malf_check, ani_check, vk_check)
